+++ src/onegov.swissvotes/HISTORY.rst
@@ -1,5 +1,9 @@
 Changelog
 ---------
+
+- Makes the update, export and delete vote views to be indifferent to the filters.
+  [msom]
+
 0.0.6 (2018-09-23)
 ~~~~~~~~~~~~~~~~~~~